refactore create_book method and create CreateBook Class
